feat: Set `display: none` on `Hidden` when its `visible` prop is set to `false`
People setting CSS display on the `Hidden` component often get confused by the fact that it stops hiding when the `visible` is `false`.

This was happening because Reakit has only set the `hidden` attribute, which gets overriden by the CSS display set by the user.

This ensures that the component will be hidden by setting `display: none` as inline style.

This shouldn't be a breaking change unless there's a weird edge case.