Update Menu with Framer Motion example (#2328)
This PR improves the `menu-framer-motion` example. The code has been
simplified, and the documentation has been enhanced to ensure clarity.

Furthermore, two bugs that were identified have been resolved. The first
bug was related to controlled props not updating the state setter, and
the second bug was related to the `Menu` not properly displaying when
set to `hidden=false`.