Use Floating UI within the component instead of store (#2279)
Closes #887
Closes #1595
Closes #2228

The primary purpose of this PR is to move the `@floating-ui/dom`
dependency from the `usePopoverStore` hook to the `Popover` component so
it can be lazy loaded.

In addition, we moved most of the Floating UI-related props to the
`Popover` component, which was the original idea, but wasn't viable
before without hacks. The only exception is the `placement` prop, to
which we need access on components other than `Popover`.

This introduces breaking changes on `Popover` and derived components.
Those are described in the changesets.

---

Finally, the `Tooltip` component has been updated, so it composes
`Hovercard` now, which also benefited from the `Popover` changes. The
specific `Tooltip` updates are related to the closed issues above.