Add `preserveTabOrderAnchor` prop to `Portal` (#3059)
This PR adds a new `preserveTabOrderAnchor` prop to `Portal` and related
components. The `Popover` component also automatically sets it as its
disclosure element.

This ensures that, when the `portal` prop is enabled, the tab order will
be preserved from the disclosure to the content element even when the
`Popover` component is rendered in a different location in the React
tree when the' portal' prop is enabled.