Fix Select not opening when unmountOnHide is true (#3151)
Closes #3147

When the select popover was mounted by default, the external/context
store was only initialized as part of the internal store, not the
host/provider component. As a result, if the popover were unmounted, it
would be destroyed, leaving the external/context store uninitiated.

This PR modifies the store logic to enable a store to be initialized
again before it's destroyed.