Render Tooltip as a visual description only (#3318)
Closes #3242

This PR removes the `aria-describedby` attribute from tooltip
components. As mentioned in #3242, this can lead to screen readers
announcing duplicate labels when the tooltip and the anchor's accessible
name share the same content. This can also become an issue when tooltips
are dynamically rendered.

However, we're not getting rid of `role="tooltip"`. The tests showed
that this role doesn't impact screen readers. Moreover, its removal
could disrupt people's code, as they might use the `role` to select the
element in their tests, for instance. If the `role` is removed, there's
no other straightforward method to select the tooltip element except by
adding specific attributes for tests (like `data-testid`).