Disable role when rendering nested modals (#3341)
This PR revises the `disableTreeOutside` logic. Now, when a modal dialog
is open and nested within another element that has a `role` attribute in
the DOM, the ancestor's `role` will be set to `"none"` until the nested
modal is closed. This change aims to prevent screen readers from moving
the virtual cursor to an ancestor while a nested modal is open.