Fix unmountOnHide on TabPanel (#3559)
When `TabPanel` has no `tabId`, it uses the element position in the DOM
to determine the related tab. If we completely unmount the tab panel
element, the active tab panel will be always rendered as the first one,
and thus connected to the first tab.

This PR updates the logic so only the panel children will be unmounted,
not the panel element itself.