Fix mousedown on MenuButton hiding Menu (#4093)
This PR addresses a long-standing issue where `mousedown` on
`MenuButton` hides the menu when it's inside a focusable ancestor (like
a dialog) on Safari.

This occurs because Safari focuses on the nearest focusable ancestor
when a button is clicked, instead of focusing on the button itself like
other browsers. Ariakit normalizes this behavior in a non-intrusive way
by restoring focus to the button, but it doesn't prevent the focusable
ancestor from receiving focus (as it would also prevent other side
effects like drag events, selection, etc.).

The problem is that allowing the focusable ancestor to receive focus
also triggers the `hideOnInteractOutside` feature on dialogs. This PR
adds a flag to the focused ancestor that's checked in the hiding logic.